+++ src/components/tabs/__spec__.js
@@ -48,6 +48,9 @@ describe('Tabs', () => {
     it('sets tabValidity to a empty immutable map', () => {
       expect(instance.state.tabValidity).toEqual(Immutable.Map());
     });
+    it('sets tabWarning to a empty immutable map', () => {
+      expect(instance.state.tabWarning).toEqual(Immutable.Map());
+    });
   });
 
   describe('componentWillMount', () => {
@@ -218,16 +221,29 @@ describe('Tabs', () => {
     });
   });
 
+  describe('changeWarning', () => {
+    beforeEach(() => {
+      instance.setState({ tabWarning: Immutable.fromJS({ 'foo': false })});
+      spyOn(instance, 'setState').and.callThrough();
+      instance.changeWarning('foo', true);
+    });
+
+    it('sets the warning state for the given tab', () => {
+      expect(instance.setState).toHaveBeenCalledWith({ tabWarning: instance.state.tabWarning.set('foo', true) });
+      expect(instance.state.tabWarning.toJS()).toEqual({ 'foo': true });
+    });
+  });
+
   describe('handleTabClick', () => {
     it('sets the state to the currently selected tabId', () => {
       spyOn(instance, 'handleTabClick');
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
 
-      expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toEqual(false);
+      expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toBeFalsy();
 
       TestUtils.Simulate.click(secondTab);
 
-      expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toEqual(true);
+      expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toBeTruthy();
     });
 
     it('sets the location', () => {
@@ -284,29 +300,29 @@ describe('Tabs', () => {
   describe('tabHeaderClasses', () => {
     it('adds a carbon-tabs__header class to the tab', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
-      expect(secondTab.classList.contains('carbon-tabs__headers__header')).toEqual(true);
+      expect(secondTab.classList.contains('carbon-tabs__headers__header')).toBeTruthy();
     });
 
     it('adds the className included in the props to the tab', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
-      expect(secondTab.classList.contains('headerClass2')).toEqual(true);
+      expect(secondTab.classList.contains('headerClass2')).toBeTruthy();
     });
 
     it('adds the className included in the props as an array to the tab', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[2];
-      expect(secondTab.classList.contains('headerClass3')).toEqual(true);
-      expect(secondTab.classList.contains('headerClass4')).toEqual(true);
+      expect(secondTab.classList.contains('headerClass3')).toBeTruthy();
+      expect(secondTab.classList.contains('headerClass4')).toBeTruthy();
     });
 
-    it('does not addsa selected class to the header', () => {
+    it('does not add a selected class to the header', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
-      expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toEqual(false);
+      expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toBeFalsy();
     });
 
     describe('when tab is selected tab', () => {
       it('adds a selected class to the header', () => {
         let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[0];
-        expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toEqual(true);
+        expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toBeTruthy();
       });
     });
 
@@ -315,6 +331,22 @@ describe('Tabs', () => {
         instance.setState({ tabValidity: Immutable.fromJS({ 'uniqueid2': false })});
         let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
         expect(secondTab.className).toEqual('carbon-tabs__headers__header headerClass2 carbon-tabs__headers__header--error');
+      });
+    });
+
+    describe('when tab has a warning', () => {
+      it('adds a warning class to the header', () => {
+        instance.setState({ tabWarning: Immutable.fromJS({ 'uniqueid2': true })});
+        let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
+        expect(secondTab.className).toEqual('carbon-tabs__headers__header headerClass2 carbon-tabs__headers__header--warning');
+      });
+
+      describe('when tab has an error as well', () => {
+        it('does not add a warning class', () => {
+          instance.setState({ tabWarning: Immutable.fromJS({ 'uniqueid2': true }), tabValidity: Immutable.fromJS({ 'uniqueid2': false }) });
+          let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
+          expect(secondTab.classList.contains('carbon-tabs__headers__header--warning')).toBeFalsy();
+        });
       });
     });
   });